requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
appnope==0.1.4
asttokens==3.0.1
certifi==2026.2.25
click==8.3.2
colorama==0.4.6
comm==0.2.3
cross-web==0.6.0
debugpy==1.8.20
decorator==5.2.1
dnspython==2.8.0
email-validator==2.3.0
exceptiongroup==1.3.1
executing==2.2.1
fastapi==0.135.3
graphql-core==3.2.8
greenlet==3.4.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
ipykernel==7.1.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.8.0
jupyter_core==5.9.1
matplotlib-inline==0.2.1
mysql-connector-python==9.6.0
nest-asyncio==1.6.0
packaging==26.1
parso==0.8.5
pexpect==4.9.0
platformdirs==4.5.1
pluggy==1.6.0
prompt_toolkit==3.0.52
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
PyMySQL==1.1.2
pytest==9.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pyzmq==27.1.0
six==1.17.0
SQLAlchemy==2.0.49
SQLAlchemy-Utils==0.42.1
stack-data==0.6.3
starlette==1.0.0
strawberry-graphql==0.315.1
tomli==2.4.0
tornado==6.5.4
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.44.0
wcwidth==0.5.3
